// types/enums.ts
export enum CardType {
  CREATURE = 'CREATURE',
  STRUCTURE = 'STRUCTURE',
  SPELL = 'SPELL',
}

export enum EntityType {
  HERO = 'HERO',
  CREATURE = 'CREATURE',
  STRUCTURE = 'STRUCTURE',
}

export enum GameStatus {
  WAITING = 'WAITING',
  SETUP = 'SETUP',
  READY = 'READY',
  PLAYING = 'PLAYING',
  COMBAT = 'COMBAT',
  FINISHED = 'FINISHED',
  CANCELLED = 'CANCELLED'
}

export enum TurnPhase {
  START = 'START',
  ENERGY = 'ENERGY',
  DRAW = 'DRAW',
  ACTION = 'ACTION',
  COMBAT = 'COMBAT',
  END = 'END'
}

export enum Zone {
  DECK = 'DECK',
  HAND = 'HAND',
  BOARD = 'BOARD',
  CHANNELING = 'CHANNELING',
  DISCARD = 'DISCARD'
}

export enum ActionType {
  PLAY_CARD = 'PLAY_CARD',
  SUMMON_CREATURE = 'SUMMON_CREATURE',
  BUILD_STRUCTURE = 'BUILD_STRUCTURE',
  CHANNEL_SPELL = 'CHANNEL_SPELL',
  ACTIVATE_SPELL = 'ACTIVATE_SPELL',
  DECLARE_ATTACK = 'DECLARE_ATTACK',
  DEFEND_WITH_HERO = 'DEFEND_WITH_HERO',
  END_TURN = 'END_TURN'
}

export enum EventType {
  GAME_STARTED = 'GAME_STARTED',
  TURN_STARTED = 'TURN_STARTED',
  CARD_DRAWN = 'CARD_DRAWN',
  ENERGY_GENERATED = 'ENERGY_GENERATED',
  CARD_PLAYED = 'CARD_PLAYED',
  CREATURE_SUMMONED = 'CREATURE_SUMMONED',
  STRUCTURE_BUILT = 'STRUCTURE_BUILT',
  SPELL_CHANNELLED = 'SPELL_CHANNELLED',
  SPELL_ACTIVATED = 'SPELL_ACTIVATED',
  ATTACK_DECLARED = 'ATTACK_DECLARED',
  DICE_ROLLED = 'DICE_ROLLED',
  DAMAGE_DEALT = 'DAMAGE_DEALT',
  CREATURE_DESTROYED = 'CREATURE_DESTROYED',
  STRUCTURE_DESTROYED = 'STRUCTURE_DESTROYED',
  HERO_DAMAGED = 'HERO_DAMAGED',
  HERO_DESTROYED = 'HERO_DESTROYED',
  TURN_ENDED = 'TURN_ENDED',
  GAME_WON = 'GAME_WON'
}

export enum EntityStatus {
  SUMMONED = 'SUMMONED',
  READY = 'READY',
  EXHAUSTED = 'EXHAUSTED',
  DAMAGED = 'DAMAGED',
  DESTROYED = 'DESTROYED'
}